requirements.txt
pyyaml
requests
dataset
python-telegram-bot
